requirements.txt
python-telegram-bot==13.15
firebase-admin==6.2.0
Flask==2.3.3
gunicorn==21.2.0        # Railway will serve via gunicorn